move ObjBase and CollectionBase to legacy. Added deprecation message on instantiation
