+++ src/easyscience/legacy/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2026 EasyScience contributors <https://github.com/easyscience>
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .collection_base import CollectionBase  # noqa: F401
+from .obj_base import ObjBase  # noqa: F401
+
+__all__ = ['CollectionBase', 'ObjBase']
